+++ packages/backend/migration/1781370641857-addHiddenWidgets.js
@@ -1,0 +1,13 @@
+export class AddHiddenWidgets1781370641857 {
+    constructor() {
+        this.name = 'AddHiddenWidgets1781370641857';
+    }
+
+    async up(queryRunner) {
+        await queryRunner.query(`ALTER TABLE "meta" ADD "hiddenWidgets" character varying(1024) array NOT NULL DEFAULT '{}'`);
+    }
+
+    async down(queryRunner) {
+        await queryRunner.query(`ALTER TABLE "meta" DROP COLUMN "hiddenWidgets"`);
+    }
+}