+++ packages/backend/src/models/Meta.ts
@@ -81,6 +81,11 @@ export class MiMeta {
 	@Column('varchar', {
 		length: 1024, array: true, default: '{}',
 	})
+	public hiddenWidgets: string[];
+
+	@Column('varchar', {
+		length: 1024, array: true, default: '{}',
+	})
 	public blockedHosts: string[];
 
 	@Column('varchar', {