+++ packages/backend/src/server/api/endpoints/admin/meta.ts
@@ -161,6 +161,13 @@ export const meta = {
 					type: 'string',
 				},
 			},
+			hiddenWidgets: {
+				type: 'array',
+				optional: false, nullable: false,
+				items: {
+					type: 'string',
+				},
+			},
 			blockedHosts: {
 				type: 'array',
 				optional: false, nullable: false,
@@ -671,6 +678,7 @@ export default class extends Endpoint<typeof meta, typeof paramDef> { // eslint-
 				cacheRemoteSensitiveFiles: instance.cacheRemoteSensitiveFiles,
 				pinnedUsers: instance.pinnedUsers,
 				hiddenTags: instance.hiddenTags,
+				hiddenWidgets: instance.hiddenWidgets,
 				blockedHosts: instance.blockedHosts,
 				silencedHosts: instance.silencedHosts,
 				mediaSilencedHosts: instance.mediaSilencedHosts,